+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+- add `sync` unit tests
+- clean `sync` code
+- remove syncing attempts on protocol startup
+- add message schema validation to inbound protocol updates
+
 ## 0.0.21
 
 - fix `iframe-app` docker image to not override env vars on start